Vista partits acabat V1

+++ PorraGirona/Controllers/PartitsController.cs
@@ -22,14 +22,36 @@ namespace PorraGirona.Controllers
         {
             _context = new PostDbContext();
         }
+        //Nivell Accés
+        private int NivellAcces()
+        {
+            int nivell = 10;
 
+            String rol;
+            byte[] valor = null;
+            bool existeix = HttpContext.Session.TryGetValue("rol", out valor);
+            if (valor != null)
+            {
+                rol = System.Text.Encoding.UTF8.GetString(valor);
+                if (rol == "admin")
+                    nivell = 0;
+                if (rol == "soci")
+                    nivell = 5;
+            }
+            return nivell;
+        }
 
         // GET: Partits
         public async Task<IActionResult> Index()
         {
-            //var postDbContext = _context.Partits.Include(p => p.IdequiplocalNavigation).Include(p => p.IdequipvisitantNavigation);
-            //return View(await postDbContext.ToListAsync());
-            return View(await _context.Partits.OrderBy(p => p.Jornada).ToListAsync());
+            if (NivellAcces() != 0) return RedirectToAction("Index","Login");
+            else
+            {
+                //var postDbContext = _context.Partits.Include(p => p.IdequiplocalNavigation).Include(p => p.IdequipvisitantNavigation);
+                //return View(await postDbContext.ToListAsync());
+                return View(await _context.Partits.OrderBy(p => p.Jornada).ToListAsync());
+            }
+
         }
 
         // GET: Partits/Details/5
@@ -39,25 +61,32 @@ namespace PorraGirona.Controllers
             {
                 return NotFound();
             }
-
-            var partit = await _context.Partits
+            if (NivellAcces() != 0) return RedirectToAction("Index", "Login");
+            else
+            {
+                var partit = await _context.Partits
                 .Include(p => p.IdequiplocalNavigation)
                 .Include(p => p.IdequipvisitantNavigation)
                 .FirstOrDefaultAsync(m => m.Idpartit == id);
-            if (partit == null)
-            {
-                return NotFound();
-            }
+                if (partit == null)
+                {
+                    return NotFound();
+                }
 
-            return View(partit);
+                return View(partit);
+            }
         }
 
         // GET: Partits/Create
         public IActionResult Create()
         {
-            ViewData["Idequiplocal"] = new SelectList(_context.Equips, "Idequip", "Idequip");
-            ViewData["Idequipvisitant"] = new SelectList(_context.Equips, "Idequip", "Idequip");
-            return View();
+            if (NivellAcces() != 0) return RedirectToAction("Index", "Login");
+            else
+            {
+                ViewData["Idequiplocal"] = new SelectList(_context.Equips, "Idequip", "Idequip");
+                ViewData["Idequipvisitant"] = new SelectList(_context.Equips, "Idequip", "Idequip");
+                return View();
+            }
         }
 
         // POST: Partits/Create
@@ -67,15 +96,19 @@ namespace PorraGirona.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Idpartit,Idequiplocal,Idequipvisitant,Datainici,Jornada,Golslocal,Golsvisitant,Finalitzat,Temporada,Idsjugadorslocal,Idsjugadorsvisitant")] Partit partit)
         {
-            if (ModelState.IsValid)
+            if (NivellAcces() != 0) return RedirectToAction("Index", "Login");
+            else
             {
-                _context.Add(partit);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(partit);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["Idequiplocal"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequiplocal);
+                ViewData["Idequipvisitant"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequipvisitant);
+                return View(partit);
             }
-            ViewData["Idequiplocal"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequiplocal);
-            ViewData["Idequipvisitant"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequipvisitant);
-            return View(partit);
         }
 
         // GET: Partits/Edit/5
@@ -86,14 +119,18 @@ namespace PorraGirona.Controllers
                 return NotFound();
             }
 
-            var partit = await _context.Partits.FindAsync(id);
-            if (partit == null)
+            if (NivellAcces() != 0) return RedirectToAction("Index", "Login");
+            else
             {
-                return NotFound();
+                var partit = await _context.Partits.FindAsync(id);
+                if (partit == null)
+                {
+                    return NotFound();
+                }
+                ViewData["Idequiplocal"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequiplocal);
+                ViewData["Idequipvisitant"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequipvisitant);
+                return View(partit);
             }
-            ViewData["Idequiplocal"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequiplocal);
-            ViewData["Idequipvisitant"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequipvisitant);
-            return View(partit);
         }
 
         // POST: Partits/Edit/5
@@ -107,30 +144,33 @@ namespace PorraGirona.Controllers
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            if (NivellAcces() != 0) return RedirectToAction("Index", "Login");
+            else
             {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(partit);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!PartitExists(partit.Idpartit))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(partit);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!PartitExists(partit.Idpartit))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["Idequiplocal"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequiplocal);
+                ViewData["Idequipvisitant"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequipvisitant);
+                return View(partit);
             }
-            ViewData["Idequiplocal"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequiplocal);
-            ViewData["Idequipvisitant"] = new SelectList(_context.Equips, "Idequip", "Idequip", partit.Idequipvisitant);
-            return View(partit);
         }
 
         // GET: Partits/Delete/5
@@ -140,17 +180,20 @@ namespace PorraGirona.Controllers
             {
                 return NotFound();
             }
-
-            var partit = await _context.Partits
+            if (NivellAcces() != 0) return RedirectToAction("Index", "Login");
+            else
+            {
+                var partit = await _context.Partits
                 .Include(p => p.IdequiplocalNavigation)
                 .Include(p => p.IdequipvisitantNavigation)
                 .FirstOrDefaultAsync(m => m.Idpartit == id);
-            if (partit == null)
-            {
-                return NotFound();
-            }
+                if (partit == null)
+                {
+                    return NotFound();
+                }
 
-            return View(partit);
+                return View(partit);
+            }
         }
 
         // POST: Partits/Delete/5
@@ -158,10 +201,14 @@ namespace PorraGirona.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var partit = await _context.Partits.FindAsync(id);
-            _context.Partits.Remove(partit);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (NivellAcces() != 0) return RedirectToAction("Index", "Login");
+            else
+            {
+                var partit = await _context.Partits.FindAsync(id);
+                _context.Partits.Remove(partit);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
         }
 
         private bool PartitExists(int id)
